requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
google-api-python-client==2.159.0
google-auth==2.37.0
google-auth-oauthlib==1.2.1
xero-python==9.5.0
anthropic==0.42.0
python-dotenv==1.0.1
pydantic-settings==2.7.1
httpx==0.28.1
tenacity==9.0.0
